js/src/xpconnect/tests/mochitest/test_bug618017.html

<!DOCTYPE html>
<html>





<head>
  <title>Test for Bug 618017</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type='application/javascript;version=1.7'>
let x = 12;
function doLetEval() {
  ok(eval('let x = 13; x') === 13, 'let statement is valid syntax in version 1.7');
}
</script>
<script type='application/javascript;version=1.5'>
ok(uneval(<test />) === '<test/>', 'xml can be parsed in any version');
doLetEval(); // Call to a function with a different version.
</script>
</body>
</html>
